+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./management/client";
 export { ClientConfig } from "./config/config";
+export * from "./util/token";